+++ drivers/cloudflare_imgbed/driver.go
@@ -1,4 +1,4 @@
-package cfimgbed
+package cloudflare_imgbed
 
 import (
 	"context"
