+++ drivers/cloudflare_imgbed/types.go
@@ -1,4 +1,4 @@
-package cfimgbed
+package cloudflare_imgbed
 
 import (
 	"fmt"
