+++ drivers/cloudflare_imgbed/util.go
@@ -1,3 +1,3 @@
-package cfimgbed
+package cloudflare_imgbed
 
 // do others that not defined in Driver interface
